Initial mini release
Main goal of governator-mini is to provide a simplified version
to the excellent governator library. To have the minimal set of
dependencies some features are removed from the original:
- no classpath scanning
- no auto binding
- no validation
- no grapher
- no Archaius bindings

Library compatibility updated to Java 7, backward compat lib
removed, guava dependency updated to 14.0.1

After all this the library only depends on guice and guava.